FEM - 1677_1
change threshhold to long

+++ kavaplugin/src/main/java/com/kaltura/playkit/plugins/kava/KavaAnalyticsConfig.java
@@ -56,7 +56,7 @@ public class KavaAnalyticsConfig {
         return this;
     }
 
-    public KavaAnalyticsConfig setDvrThreshold(int dvrThreshold) {
+    public KavaAnalyticsConfig setDvrThreshold(long dvrThreshold) {
         this.dvrThreshold = dvrThreshold;
         return this;
     }
